definicion de roles y panel de administracion (faltan elementos))

+++ app/src/main/java/com/example/nextbyte_app/data/ProductData.kt
@@ -4,7 +4,7 @@ data class Product(
     val id: String = "",  // IMPORTANTE: agregar ID para Firebase
     val name: String = "",
     val description: String = "",
-    val price: Double = 0.0,
+    val price: Int = 0,
     val imageUrl: String = "",
     val category: String = "",
     val stock: Int = 0,
@@ -12,5 +12,5 @@ data class Product(
     val isFavorited: Boolean = false
 ) {
     // Constructor sin parámetros para Firebase
-    constructor() : this("", "", "", 0.0, "", "", 0, 0.0)
+    constructor() : this("", "", "", 0, "", "", 0, 0.0)
 }
